Update aStar.cpp
adds the sleep functionality to the program. This is done so that a user who has chosen have their view updated every move or few moves, will be able to get a better idea of what moves the algorithm is taking, not a perfect solution in this case due to how the console looks.